new try

+++ hashmap.c
@@ -118,7 +118,7 @@ Pair * searchMap(HashMap * map,  char * key) {
 }
 
 Pair * firstMap(HashMap * map) {
-    for (int i=0; i < map -> capacity; i++){
+    for (long i=0; i < map -> capacity; i++){
         if((map -> buckets[i]) && (map -> buckets[i+1] -> key) && (map -> buckets[i+1] -> value)){
             map -> current = i;
             return map -> buckets[i];
@@ -128,7 +128,7 @@ Pair * firstMap(HashMap * map) {
 }
 
 Pair * nextMap(HashMap * map) {
-for (int i = map -> current ; i < map -> capacity; i++){
+for (long i = map -> current ; i < map -> capacity; i++){
     if((map -> buckets[i+1]) && (map -> buckets[i+1] -> key) && (map -> buckets[i+1] -> value )){
         map -> current = (i+1);
         return map -> buckets[i+1];
